fix: Fix linter by ignoring PascalCase
In this context, I want `home` to refer to the controller file name,
which is in lower case because it is not refering to a class name. I
might want to reconsider this approach in the future to match better
with PHP standards, but this is not the case today.

I don't set the rule in a configuration file because I prefer to make
exceptions explicit.