tec: Fix test for missing amount
By generating more data (see previous commit), I discovered the test
testing amount between 0 and 0.99€ was failing with a value of 0. It is
indeed considered as "no amount" and the message wasn't the expected
one.

In fact, there are no needs to make a distinction between no amounts
(i.e. `null`) and an amount of 0€: there are the same case.

I've set the default value of `amount` to 0 to make it clear.