tec: Remove model constructors
It appeared that the models were all created the same way (with a
PROPERTIES and a constructor passing this constant to the parent
constructor). Properties were duplicated in memory which could become a
problem with a lot of models.

I made the PROPERTIES const to be loaded automatically for the classes
inherating from Model and store them in a static variable. It saves few
lines and memory.