Save preferred frequency on renew
I directly changed an old migration in this commit: you must NEVER do
that. Because the migration is on master and is theorically already
applied in production, it will not be rerun and so the two columns will
not have their "defaulted not null" configuration. I did it because:

- I'm working alone
- I know the migration is not applied yet in production nor anywhere
  else
- I'm able to apply the change easily myself on my development DB
- "ALTER TABLE" is extremely painful with SQLite
- I'm aware I had to do differently and I only have bad excuses to do it
  this way
- this commit probably took me more time that it would have to write a
  proper migration

So remember: NEVER change an old migration!